third_party/blink/web_tests/fast/dom/elementsFromPoint/elementsFromPoint-shadowroot-expected.txt

 
PASS shadowRoot.elementsFromPoint() threw exception TypeError: Failed to execute 'elementsFromPoint' on 'ShadowRoot': 2 arguments required, but only 0 present..
PASS shadowRoot.elementsFromPoint(0) threw exception TypeError: Failed to execute 'elementsFromPoint' on 'ShadowRoot': 2 arguments required, but only 1 present..
PASS nodeListToString(shadowRoot.elementsFromPoint(x12, y12)) is "DIV, DIV#host, BODY, HTML"
PASS nodeListToString(nestedShadowRoot.elementsFromPoint(x22, y22)) is "DIV, DIV, DIV#host, BODY, HTML"
PASS nodeListToString(shadowRoot.elementsFromPoint(x22, y22)) is "DIV, DIV#host, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(x22, y22)) is "DIV#host, BODY, HTML"
PASS nodeListToString(root3.elementsFromPoint(centerX(blockHost), centerY(blockHost))) is "DIV#blockHost, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(centerX(blockHost), centerY(blockHost))) is "DIV#blockHost, BODY, HTML"
PASS nodeListToString(root4.elementsFromPoint(centerX(inlineBlockHost), centerY(inlineBlockHost))) is "SPAN#inlineBlockHost, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(centerX(inlineBlockHost), centerY(inlineBlockHost))) is "SPAN#inlineBlockHost, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(centerX(submit), centerY(submit))) is "INPUT#submit, BODY, HTML"
PASS successfullyParsed is true

TEST COMPLETE
